Change calc_utm_items score param to float

+++ report_utils.py
@@ -8,7 +8,7 @@ import generate_topology
 from common_constants import DANGER_COUNTRIES
 
 
-def calc_utm_items(score: int, open_ports: Iterable[str], countries: Iterable[str]) -> List[str]:
+def calc_utm_items(score: float, open_ports: Iterable[str], countries: Iterable[str]) -> List[str]:
     """Return which UTM features help mitigate the detected risks."""
     items = set()
     if list(open_ports):
